AbstractSynchronisationService
